Fix relay acceptance checks, URL validation, and state management issues
Key fixes:
- Add relay acceptance validation to music_player, webbookmarks, bookmarks,
  share_modal, live/share_modal, git_hosting modules, recipe_store,
  media.rs, topic_store/publish, and dvm_store
- Fix relay acceptance helper to include per-relay failure details
- Add URL scheme validation for image URLs in ai_chat rendering
- Fix relay validation in zapgoals to return error on invalid relay URLs
- Fix published_event_id short-circuit to update retry metadata
- Include local_pending_ tokens in cashu reconciliation
- Gate new_tokens push on queue success in cashu/mpp
- Fix DVM_LAST_REQUEST_ID not cleared on failure
- Move HAS_SIGNER check before ensure_relays_ready in muting
- Add retry button to podcast RSS detail error view
- Fix format_bytes to use u64 instead of usize
- Fix bg-linear-to-r → bg-gradient-to-r and bg-linear-to-br → bg-gradient-to-br

+++ src/utils/format.rs
@@ -24,7 +24,7 @@ pub fn display_server_url(url: &str) -> String {
 }
 
 /// Format bytes into a human-readable binary size string.
-pub fn format_bytes(bytes: usize) -> String {
+pub fn format_bytes(bytes: u64) -> String {
     let units = ["B", "KB", "MB", "GB", "TB", "PB", "EB"];
     let mut value = bytes as f64;
     let mut unit_index = 0usize;
